bitcoin_transactions_version_1
The current version contains parser for monitoring bitcoin transactions. It can be used for analyzing suspicious activity or extra massive transactions. Of course, script can be developed. Here it is possible to monitor not only bitcoin, but many others cryptos and provide a choise for user which crypto to monitor, but I decided to code as simple as possible. Best wishes...